Fixed Bug #803
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@9992 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/ports/STM32/LLD/DACv1/hal_dac_lld.h
@@ -374,7 +374,7 @@ typedef struct {
   /**
    * @brief   DAC control register.
    */
-  uint32_t                  cr;
+  uint16_t                  cr;
 } DACConfig;
 
 /**
